requirements.txt
Django==2.0.5
django-bootstrap3==9.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser